*: revert to toolkit version 3.2.109
In toolkit version 3.2.131 we introduced a regression where
we are untarring cache contents with --skip-old-file. This means
that if .cargo/cargo.toml exists before the cache restore, we will
skip restoring the most up to date version of this toml file in
the cache action. We temporarily move back to 3.2.109 while
we fix this regression.